model/canavar-model/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data
pyarrow>=13.0.0
requests>=2.31.0

# Experiment management
pyyaml>=6.0
wandb>=0.15.0
tqdm>=4.65.0

# Federated Learning
flwr>=1.5.0
flwr[simulation]>=1.5.0

# Evaluation & Explainability
shap>=0.42.0
umap-learn>=0.5.3
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Adversarial robustness
advertorch>=0.2.4

# Utilities
python-dotenv>=1.0.0
